\name{conctab}
\alias{conctab}
\docType{data}
\title{
Concentrations of drugs used in the drug screen
}
\description{
This data set contains drug concentrations used in the drug screen.
Each of the 64 compounds (drug IDs as row names) was screened in 5
concentrations steps c1-c5 (column names). The column 'c1' indicates
the highest, and 'c5' indicates the lowest drug concentration used
in the screen.
}
\author{Małgorzata Oleś <dr.malgorzata.oles@gmail.com>}
\usage{
conctab
}
\format{
data.frame with 64 rows and 5 columns.
}
\keyword{datasets}
