\name{drugs}
\alias{drugs}
\docType{data}
\title{
Meta data of the compounds
}
\description{
This data set contains additional information about the drugs used in the
drug screens. Row names contain drug IDs. The data.frame
contains 8 columns, which provide information on: the official drug
name ('name'), main targets ('main_targets'), target category
('target_category'), pathway annotation ('group', 'pathway'), distributor,
and whether the drug was approved ('approved_042016') or was in the development
stage ('devel_042016').
}
\author{Małgorzata Oleś <dr.malgorzata.oles@gmail.com>}
\usage{
drugs
}
\format{
data.frame with 91 rows and 8 columns.
}
\keyword{datasets}
