\name{exp10div}
\alias{exp10div}
\title{
Axis labels for p-values
}
\description{
The function formats axis labels of p-values in a nice way.
}
\usage{
exp10div(x)
}
\arguments{
  \item{x}{numeric}
}
\value{
Object of class expression
}
\author{
Małgorzata Oleś <malgorzata.oles@embl.de>
}
\examples{
exp10div(-10)
}
