\name{percentAxisScale}
\alias{percentAxisScale}
\title{
Fraction to percent converter
}
\description{
The function converts fractions to percent by multiplying input value by 100.
}
\usage{
percentAxisScale(x)
}
\arguments{
  \item{x}{numeric vector}
}
\value{
numeric vector
}
\author{
Małgorzata Oleś <malgorzata.oles@embl.de>
}
\examples{
percentAxisScale(x=c(0, 0.1, 1))
}
