\name{scientific_10}
\alias{scientific_10}
\title{
log10 scale labels in ggplot2
}
\description{
This function is useful for formatting labels in ggplot2 of log10 axis.
}
\usage{
scientific_10
}
\value{
numeric vector
}
\author{
Małgorzata Oleś <malgorzata.oles@embl.de>
}
\examples{
## scale_x_log10(labels=scientific_10)
}
